# toolchain/spec/openapi.yaml
components:
...
  schemas:
...
    Attachment:
      additionalProperties: false
      properties:
        mime_type:
          type: string
        url:
          $ref: '#/components/schemas/URL'
      required:
      - url
      - mime_type
      type: object
...
    Bf16QuantizationConfig:
      additionalProperties: false
      properties:
        quantization_type:
          const: bf16
          type: string
      required:
      - quantization_type
      type: object
...
    ChatCompletionRequest:
      additionalProperties: false
      properties:
        available_tools:
          items:
            $ref: '#/components/schemas/ToolDefinition'
          type: array
        logprobs:
          additionalProperties: false
          properties:
            top_k:
              type: integer
          type: object
        messages:
          items:
            oneOf:
            - $ref: '#/components/schemas/UserMessage'
            - $ref: '#/components/schemas/SystemMessage'
            - $ref: '#/components/schemas/ToolResponseMessage'
            - $ref: '#/components/schemas/CompletionMessage'
          type: array
        model:
          $ref: '#/components/schemas/InstructModel'
        quantization_config:
          oneOf:
          - $ref: '#/components/schemas/Bf16QuantizationConfig'
          - $ref: '#/components/schemas/Fp8QuantizationConfig'
        sampling_params:
          $ref: '#/components/schemas/SamplingParams'
        stream:
          type: boolean
      required:
      - model
      - messages
      type: object
...
    ChatCompletionResponseEvent:
      additionalProperties: false
      properties:
        delta:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ToolCallDelta'
        event_type:
          $ref: '#/components/schemas/ChatCompletionResponseEventType'
        logprobs:
          items:
            $ref: '#/components/schemas/TokenLogProbs'
          type: array
        stop_reason:
          $ref: '#/components/schemas/StopReason'
      required:
      - event_type
      - delta
      title: Chat completion response event.
      type: object
    ChatCompletionResponseEventType:
      enum:
      - start
      - complete
      - progress
      type: string
    ChatCompletionResponseStreamChunk:
      additionalProperties: false
      properties:
        event:
          $ref: '#/components/schemas/ChatCompletionResponseEvent'
      required:
      - event
      title: SSE-stream of these events.
      type: object
    CompletionMessage:
      additionalProperties: false
      properties:
        content:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/Attachment'
          - items:
              oneOf:
              - type: string
              - $ref: '#/components/schemas/Attachment'
            type: array
        role:
          const: assistant
          type: string
        stop_reason:
          $ref: '#/components/schemas/StopReason'
        tool_calls:
          items:
            $ref: '#/components/schemas/ToolCall'
          type: array
      required:
      - role
      - content
      - stop_reason
      - tool_calls
      type: object
...
    Fp8QuantizationConfig:
      additionalProperties: false
      properties:
        quantization_type:
          const: fp8
          type: string
      required:
      - quantization_type
      type: object
    InstructModel:
      enum:
      - llama3_8b_chat
      - llama3_70b_chat
      type: string
...
    SamplingParams:
      additionalProperties: false
      properties:
        max_tokens:
          type: integer
        repetition_penalty:
          type: number
        strategy:
          $ref: '#/components/schemas/SamplingStrategy'
        temperature:
          type: number
        top_k:
          type: integer
        top_p:
          type: number
      required:
      - strategy
      type: object
    SamplingStrategy:
      enum:
      - greedy
      - top_p
      - top_k
      type: string
...
    StopReason:
      enum:
      - end_of_turn
      - end_of_message
      - out_of_tokens
      type: string
...
    SystemMessage:
      additionalProperties: false
      properties:
        content:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/Attachment'
          - items:
              oneOf:
              - type: string
              - $ref: '#/components/schemas/Attachment'
            type: array
        role:
          const: system
          type: string
      required:
      - role
      - content
      type: object
    TokenLogProbs:
      additionalProperties: false
      properties:
        logprobs_by_token:
          additionalProperties:
            type: number
          type: object
      required:
      - logprobs_by_token
      type: object
    ToolCall:
      additionalProperties: false
      properties:
        arguments:
          additionalProperties:
            oneOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
            - items:
                oneOf:
                - type: string
                - type: integer
                - type: number
                - type: boolean
                - type: 'null'
              type: array
            - additionalProperties:
                oneOf:
                - type: string
                - type: integer
                - type: number
                - type: boolean
                - type: 'null'
              type: object
          type: object
        call_id:
          type: string
        tool_name:
          oneOf:
          - enum:
            - brave_search
            - wolfram_alpha
            - photogen
            - code_interpreter
            type: string
          - type: string
      required:
      - call_id
      - tool_name
      - arguments
      type: object
    ToolCallDelta:
      additionalProperties: false
      properties:
        content:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ToolCall'
        parse_status:
          $ref: '#/components/schemas/ToolCallParseStatus'
      required:
      - content
      - parse_status
      type: object
    ToolCallParseStatus:
      enum:
      - start
      - in_progress
      - failure
      - success
      type: string
    ToolDefinition:
      additionalProperties: false
      properties:
        description:
          type: string
        parameters:
          additionalProperties:
            $ref: '#/components/schemas/ToolParamDefinition'
          type: object
        tool_name:
          oneOf:
          - enum:
            - brave_search
            - wolfram_alpha
            - photogen
            - code_interpreter
            type: string
          - type: string
      required:
      - tool_name
      type: object
...
    ToolParamDefinition:
      additionalProperties: false
      properties:
        description:
          type: string
        param_type:
          type: string
        required:
          type: boolean
      required:
      - param_type
      type: object
...
    ToolResponseMessage:
      additionalProperties: false
      properties:
        call_id:
          type: string
        content:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/Attachment'
          - items:
              oneOf:
              - type: string
              - $ref: '#/components/schemas/Attachment'
            type: array
        role:
          const: ipython
          type: string
        tool_name:
          oneOf:
          - enum:
            - brave_search
            - wolfram_alpha
            - photogen
            - code_interpreter
            type: string
          - type: string
      required:
      - role
      - call_id
      - tool_name
      - content
      type: object
...
    URL:
      format: uri
      pattern: ^(https?://|file://|data:)
      type: string
    UserMessage:
      additionalProperties: false
      properties:
        content:
          oneOf:
          - type: string
          - $ref: '#/components/schemas/Attachment'
          - items:
              oneOf:
              - type: string
              - $ref: '#/components/schemas/Attachment'
            type: array
        role:
          const: user
          type: string
      required:
      - role
      - content
      type: object
info:
  description: "This is the specification of the llama stack that provides\n     \
    \           a set of endpoints and their corresponding interfaces that are tailored\
    \ to\n                best leverage Llama Models. The specification is still in\
    \ draft and subject to change.\n                Generated at 2024-07-19 11:49:56.794897"
  title: '[DRAFT] Llama Stack Specification'
  version: 0.0.1
...
openapi: 3.1.0
paths:
...
  /inference/chat_completion:
    post:
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponseStreamChunk'
          description: SSE-stream of these events.
      tags:
      - ModelInference
...
security:
- Default: []
servers:
- url: http://any-hosted-llama-stack.com
...